refactor(grid): add checkbox plugin
