+++ packages/vue/src/grid/index.ts
@@ -32,6 +32,7 @@ import columnAnchor from './src/column-anchor'
 import dragger from './src/dragger'
 import sort from './src/sort'
 import tooltip from './src/tooltip'
+import checkbox from './src/checkbox'
 import * as GridTools from './src/tools'
 import { version } from './package.json'
 import type { Plugin } from './src/types/index.type'
@@ -51,6 +52,7 @@ import type { Plugin } from './src/types/index.type'
  * dragger 拖拽相关逻辑
  * sort 排序相关逻辑
  * tooltip 提示相关逻辑
+ * checkbox 多选相关逻辑
  */
 const plugins: Plugin[] = [
   Menu,
@@ -66,7 +68,8 @@ const plugins: Plugin[] = [
   columnAnchor,
   dragger,
   sort,
-  tooltip
+  tooltip,
+  checkbox
 ]
 
 // 设置全局参数,配置GlobalConfig，提供比如国际化方法
